fix: 일반 사용자는 테마를 추가, 삭제할 수 없음
