+++ src/main/resources/static/js/theme.js
@@ -1,5 +1,6 @@
 let isEditing = false;
-const API_ENDPOINT = '/themes';
+const ADMIN_API_ENDPOINT = '/admin/themes';
+const USER_API_ENDPOINT = '/themes';
 const cellFields = ['id', 'name', 'description', 'thumbnail'];
 const createCellFields = ['', createInput(), createInput(), createInput()];
 function createBody(inputs) {
@@ -108,7 +109,7 @@ function requestCreate(data) {
     body: JSON.stringify(data)
   };
 
-  return fetch(API_ENDPOINT, requestOptions)
+  return fetch(ADMIN_API_ENDPOINT, requestOptions)
       .then(response => {
         if (response.status === 201) return response.json();
         throw new Error('Create failed');
@@ -116,7 +117,7 @@ function requestCreate(data) {
 }
 
 function requestRead() {
-  return fetch(API_ENDPOINT)
+  return fetch(USER_API_ENDPOINT)
       .then(response => {
         if (response.status === 200) return response.json();
         throw new Error('Read failed');
@@ -128,7 +129,7 @@ function requestDelete(id) {
     method: 'DELETE',
   };
 
-  return fetch(`${API_ENDPOINT}/${id}`, requestOptions)
+  return fetch(`${ADMIN_API_ENDPOINT}/${id}`, requestOptions)
       .then(response => {
         if (response.status !== 204) throw new Error('Delete failed');
       });
